TC refactor
