+++ apps/backend/question/src/models/Question.ts
@@ -1,4 +1,7 @@
 import mongoose, { Document, Schema } from "mongoose";
+import moment from "moment-timezone";
+
+// console.log("time:" , moment().tz("Asia/Singapore").format())
 
 interface Question extends Document {
   id: string;
@@ -54,24 +57,30 @@ const questionSchema = new Schema({
   ],
   dateCreated: {
     type: Date,
-    default: Date.now, // Set the default value to the current date and time
+    // Dysfunctional
+    default: function() {
+      console.log("date:", new Date(moment().tz("Asia/Singapore").format()))
+      return new Date().toLocaleString("en-SG", {timeZone: "Asia/Singapore"});
+    }, // Set the default value to the current date and time
     required: false,
-  },
-});
+  }, 
+},  
+{versionKey: false} // Removes the __v: 0 attribute
+);
+
 
 // Removes the __v: 0 attribute 
 questionSchema.set('toJSON', {
   transform: (document, returnedObject) => {
       returnedObject.id = returnedObject._id.toString()
-      delete returnedObject.__v
-
-    // Remove _id from the testcases array
-    if (Array.isArray(returnedObject.testcases)) {
-      returnedObject.testcases = returnedObject.testcases.map((tc) => {
-        const { _id, ...rest } = tc;
-        return rest;
-      });
-    }
+    // // Remove _id from the testcases array
+    // if (Array.isArray(returnedObject.testcases)) {
+    //   returnedObject.testcases = returnedObject.testcases.map((tc) => {
+    //     const { _id, ...rest } = tc;
+    //     delete tc.__v;
+    //     return rest;
+    //   });
+    // }
   }
 })
