+++ apps/frontend/src/components/questions/questionPage/codeEditor/ExecPanel/ResultContent.tsx
